+++ lib/classes/classes.dart
@@ -1,0 +1,22 @@
+
+class Property {
+  String title;
+  String imageUrl;
+  int price;
+
+  Property(this.title, this.imageUrl, this.price);
+}
+
+class Tenant {
+  String name;
+  String email;
+
+  Tenant(this.name, this.email);
+}
+
+class Agent {
+  String name;
+  String email;
+
+  Agent(this.name, this.email);
+}
